fix: remove unnecessary logic
