+++ src/components/ErrorBoundary/WidgetErrorBoundary.module.scss
