+++ src/pages/MainPage/MainPage.tsx
@@ -2,6 +2,7 @@ import React, { ComponentProps } from 'react';
 import classes from './MainPage.module.scss';
 import { TMDBMovieSearcher } from '../../components/TMDBMovieSearcher/TMDBMovieSearcher';
 import { PageWrapper } from '../../components/PageWrapper/PageWrapper';
+import { WidgetErrorBoundary } from '../../components/ErrorBoundary/WidgetErrorBoundary';
 
 interface MainPageProps extends ComponentProps<'div'> {}
 
@@ -11,7 +12,9 @@ export class MainPage extends React.Component<MainPageProps> {
       <main {...this.props} className={classes.page}>
         <PageWrapper>
           <h1 className="visually-hidden">React Component Task</h1>
-          <TMDBMovieSearcher />
+          <WidgetErrorBoundary>
+            <TMDBMovieSearcher />
+          </WidgetErrorBoundary>
         </PageWrapper>
       </main>
     );
